Merge: warn on wrong feature/hotfix command
Now you'll have to confirm if you want to hotfix-merge a feature, or
feature-merge a hotfix.  This should reduces silly mistakes.

`current_hotfix_branch` no longer exits if we're not on a hotfix.  It was
strange that we did that sort of checking for hotfixes, but not features, so
now we should implement that sort of constraint, where appropriate, in the
calling functions.  This is sorta reverting part of 5ee8064d.

Inspired by #114.