Gemspec specifies exact version numbers for dependencies
Things were breaking in git-scripts because the version numbers for a
lot of dependencies were not specified and ended up being very different than
what was expected.

Also, we have two versions of Ruby floating around for now so I removed the
version specifier in the Gemfile. We can't specify "at least" 1.9.3, eventually
we want to set it to 2.0.0 since the Fedora machines will be running 2.0.0.