Add delete_current_branch function
I need this for the ifixit-scripts repo, in which I want to cancel
creation of an issue and associated branch if no issue description is
provided.

If we don't clean up the branch that we were on, when you retry
`feature start` to fix, it claims that branch name already exists.

I also tried doing the branch creation after issue creation, but that
made you go through labeling everything for nothing if the branch name
you chose was already taken.