feature prune: prune remote branches first
It's possible to have some origin/{branches} in your refs that don't
exist in the remote. If you try to delete them from the remote (cause
you think they are there), it wil fail. Let's prune our local copy of
the remote to match the origin, *before* we prune the real origin to
prevent this.